Copy the list of ETags before sorting
Some S3 multipart upload implementations (for example, the new hadoop
S3 code) pass an immutable list of ETags to `complete`. This can
cause problems when we try to sort the ETags for inclusion in the
actual request.

We can sidestep that by just copying the list before we try to sort
it.