feat: Enforce Rotation Functionality for ReEncrypt-feature  (#474)
Implemented a stronger level of validation for reEncryptInstructionFile where, when enabled, the client will attempt to decrypt the re-encrypted instruction file with the old key material and throw an exception when decryption succeeds. This is a stronger level of validation that the wrapping key has been rotated than the standard assertion that the materials descriptions are different.

---------

Co-authored-by: Anirav Kareddy <aniravk@amazon.com>